requirements.txt
Django==3.1.3
django-environ==0.4.5
djangorestframework==3.12.1
Pillow==8.0.1
psycopg2==2.8.6
python-dateutil==2.8.1